Changed rentAmount to double

+++ src/main/java/com/example/rentalmanagerapp/Rental/RentalRequest.java
@@ -11,6 +11,6 @@ public class RentalRequest {
     private String rentalAddress;
     private String description;
     private String type;
-    private long rentAmount;
+    private double rentAmount;
     private String dateAvailable;
 }
